Separate styles for all components by different files
