+++ LabProject/src/components/common/input/Input.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { Input as RNEInput } from 'react-native-elements';
-import theme from '../../theme/theme';
-import { NativeSyntheticEvent, StyleSheet, TextInputChangeEventData } from 'react-native';
+import theme from '../../../theme/theme';
+import { NativeSyntheticEvent, TextInputChangeEventData } from 'react-native';
+import styles from './styles';
 
 export interface IInputProps {
     value: string;
@@ -26,22 +27,5 @@ const Input: React.FC<IInputProps> = ({ placeholder, onChange, value, label, isP
         errorStyle={styles.errorStyle}
     />
 );
-
-const styles = StyleSheet.create({
-    container: {
-        paddingHorizontal: 0,
-    },
-    label: {
-        fontSize: 22,
-        fontWeight: '500',
-        color: theme.colors?.grey1,
-    },
-    errorStyle: {
-        fontSize: 16,
-        color: '#c20404',
-        marginTop: 10,
-        marginHorizontal: 0,
-    },
-});
 
 export default Input;
