+++ LabProject/src/components/sign-in/SignIn.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { NativeSyntheticEvent, StyleSheet, TextInputChangeEventData, View } from 'react-native';
+import { NativeSyntheticEvent, TextInputChangeEventData, View } from 'react-native';
 import { Text } from 'react-native-elements';
-import theme from '../../theme/theme';
-import Input from '../common/Input';
-import Button, { ButtonColor } from '../common/Button';
-import Chip, { ChipColor } from '../common/Chip';
+import Input from '../common/input/Input';
+import Button, { ButtonColor } from '../common/button/Button';
+import Chip, { ChipColor } from '../common/chip/Chip';
 import { SignInState } from '../../constants';
 import Icon from 'react-native-vector-icons/Ionicons';
+import styles from './styles';
 
 export interface ISignInProps {
     email: string;
@@ -89,65 +89,5 @@ const SignIn: React.FC<ISignInProps> = ({
         </View>
     );
 };
-
-const styles = StyleSheet.create({
-    root: {
-        flex: 1,
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-        flexDirection: 'column',
-        marginHorizontal: 30,
-        marginVertical: 50,
-        position: 'relative',
-    },
-    headerContainer: {
-        width: '100%',
-        flex: 0.5,
-        justifyContent: 'center',
-    },
-    mainTitleContainer: {
-        alignSelf: 'flex-start',
-        borderBottomWidth: 5,
-        borderBottomColor: theme.colors?.primary,
-    },
-    mainTitle: {
-        fontWeight: 'bold',
-        borderBottomWidth: 1,
-    },
-    inputsContainer: {
-        width: '100%',
-        flex: 2,
-    },
-    inputContainer: {
-        marginTop: 20,
-    },
-    forgotPasswordTooltip: {
-        alignSelf: 'flex-end',
-        fontSize: 16,
-        color: theme.colors?.primary,
-        fontWeight: 'bold',
-        textTransform: 'uppercase',
-        marginTop: 10,
-    },
-    footerButtonsContainer: {
-        width: '100%',
-        flex: 1,
-    },
-    loginHelpText: {
-        alignSelf: 'center',
-        fontSize: 16,
-        color: theme.colors?.grey1,
-        marginBottom: 30,
-        marginTop: 'auto',
-    },
-    pillsContainer: {
-        display: 'flex',
-        flexDirection: 'row',
-        width: '100%',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-    },
-});
 
 export default SignIn;
