+++ LabProject/src/components/tab-bar/styles.ts
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    tabBar: {
+        borderTopColor: 'black',
+        backgroundColor: 'transparent',
+    },
+    icon: {
+        width: 35,
+        height: 35,
+    },
+    label: {
+        fontSize: 12,
+        color: '#000',
+    },
+});
+
+export default styles;
