+++ LabProject/src/components/top-bar/HeaderLeft.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import Icon from 'react-native-vector-icons/Ionicons';
-import { Image, StyleSheet, TouchableOpacity } from 'react-native';
+import { Image, TouchableOpacity } from 'react-native';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import theme from '../../theme/theme';
+import { headerLeftStyles as styles } from './styles';
 
 export interface IHeaderLeftProps extends NativeStackScreenProps<Record<string, object | undefined>> {
     canGoBack: boolean;
@@ -23,16 +23,5 @@ const HeaderLeft: React.FC<IHeaderLeftProps> = ({ canGoBack, navigation }) => {
         <Icon style={styles.menuIcon} name="menu-outline" />
     );
 };
-
-const styles = StyleSheet.create({
-    menuIcon: {
-        color: theme.colors?.white,
-        fontSize: 32,
-    },
-    backIcon: {
-        width: 20,
-        height: 20,
-    },
-});
 
 export default HeaderLeft;
